make fallback to file handles more permenant. This should help keep stats for inactive mounts
as fallback files were getting setup and dropped on the fly, any mountpoints that were inactive
but had fallback to files configured where getting dropped.  To keep to a standard way of doing
things and because slaves use these stats then we should keep them around but to do so we need
to keep those file handles around by dropping the expire time to something invalid and skip them
during the purge.  They do get purged and recreated (if defined) over reload.