wildcard in IP/agent files fixed
the wildcard handling was broken in a number of cases when you placed multiple wildcards
in the files, affecting the items were checked.

A simple solution could be to use a list but for IP checking that could mean a long search
especially in IP spamming cases that dynamically add to the banned IP list. So instead I
now use 2 setups, a list for wildcards (will not be many of those) and the tree for literal
entries so the order is maintained.

It looks like there was a possible crash bug in the automatic banned IP filter as well but
as it's very rarely used probably means it hasn't been identified.