fix possible misalignment of icy metadata if headers and data are received at once
The block containing that data was getting dropped after the header processing so
ignored the initial data being sent and there fore the interval being counted for
the icy metadata block was incorrect. A behavioural change in some sendin clients
could trigger this, normally headers are sent first then data in separate packets.