avoid possible race with a stale stats handle
It could be possible for a new source to register and acquire a stats handle
after a previous source has been removed from the tree but not yet released
the stats.  In such cases, the old source would eventually release the stats
while the running source has its handle invalidated.