allow for moving a lagging listener to the front of the queue
In cases where links can have large periodic stalls, eg wireless, there can be a problem of
dropping off the queue for lagging or lagging and then doing a lot of catch-up when the link
resumes.  To ease the case of excessive sending, we now detect when the lagging listener is
about to drop off and then move them to the front of the queue. This avoids the large bursts
from the queue when resuming.

In most cases the player will go into a "buffering" phase which stalls playback, and if we
do not prevent some content from being sent then all a catch up will do is fill up the TCP
buffers at both the listener and server end.

To allow for the previous behavior, just pass norequeue in the URL parameters. It does not
require any particular value for now.