remove possible abort trigger when min queue size is reduced
the queue prune routine was had an abort trap if it ever came across the min queue. As it is
not counted, it is possible that the data could go. Normally this does not arrive but if the
min queue (or burst size) is reduced before new data is added then the prune would abort the
process. Now we just move to the next block along the queue until the prune is complete