two timing changes for source/listener client rescheduling and a automatic queue block size adjustment
With dealing with very high bitrate streams, 1 mbit+, certain internal limits prevent
sufficient bitrate from being maintained.  Most typically because information about the
rate provided is not available to certain areas without placing searches in hot paths.
The most obviously trigger here is the incoming bitrate stat which is mainatined in the
source structure but is not available to the backend writers.

use a per-client adjustment value, passed from the source to the listener to help determine
a more appropriate retry time based on the incoming bitrate.

Provide an extra reschule hook in the worker. We still have the normal schedule timers
but now also provide an optional flag that the listener can use to refer to a source
flag, this means that the source, when it has processed something, can implicitly
trigger any waiting (non-lagging) listeners to process without them waiting until some
scheduled point. This is currently used for listeners who are at the front of the queue
so are just guessing when to reschedule.

Because there are limiters in the send routines to prevent excessive amounts being sent
or recevied with one client, that together with anti busy loop triggers can limit the
max throughput of a client.  While per-client triggers are scalable the second element
is the block size read/sent.  We now scale blocks based on the incoming bitrate to at
most ~9k instead of the 1.5k that is fixed or set by the qblock size xml setting.