update to non-ogg speed and buffer size determination rules
This makes the default queue block size slightly larger (setting settable with qblock-size)
but allows the writer routines to have their own limit which is about 1400 by default. This
reduces the internal admin overhead on large queues but still imposes a packet size limiter
send to the listeners.  The only external difference is that instead of packets of N frames
going to the listener, you will get something in the order of 1400 (near common MTU).  The
setting that is new is max-send-size and is only required if you want a different write size
to 1400, but you will only ever really get up to the qblock-size, which is another reason
it was increased (currently about 2900).

There were a couple of other issues identified which were fixed but are strickly speaking
separate from the block sizing.  One is the period update of the full metadata, it did not
actually break the spec but was caused by a state setting if an EAGAIN was returned at a
specific point. I doubt that anyone noticed as it is rare and would not be seen generally.

The other aspect was the requeue was being triggerred in cases it should not of. I suspect
this could impose some performance penalty on certain setups and incorrectly trigger the
listener to skip data.