improve the range and HEAD handling.
certain players like iphone issue range requests but react poorly if a 206 response code
is returned.  To get around this we return a 200 response code (like we did until recently
but also impose a very short termination trigger.   There is also a tie-in with handling
requests using the HEAD method as thes are used to get information but not return content.

A further issue with these short duration requests is how they react when auth is in use,
which we allow to bypass the auth trigger but leave the flag disabled allowing the backend
file and stream routines to detect and drop.

The are a few changes here just because the discon time is made as a union to also allow
for a end byte position.