add chunked and keep alive capabilities to connections.
chunked response is now allowed but can be disabled via _hdr=8, but it also seems that
some players may have trouble even if HTTP 1.1 is supplied but these tend to send the
icy-metadata header so we also use that to disable chunked.

keep-alive is only enabled for 1.1 connections and clients are reset and re-added back
onto a worker for subsequent processing. A couple of the responses can be 1.1.  This
should only really be a factor for web style requests as they tend to blast a series of
requests to avoid the accept overhead.