redo a patch from the 11th for client cleanup from fallback file
The original issue still needed addressing, the fh would get destroyed if flagged with
FS_DELETE, so subsequent tests on fh would be invalid. S onow it is pushed later but not
as late as the patch on the 11th made it as the auth listener release could destroy the
client and therefore invalidate it before removal.  The multiple callers approach should
suffice making sure we have a copy of the mount string for auth if needed.