possible memory corruption fix with missing or small fallback file
if listtener gets onto the source but the source terminates before even the headers
are made then the move could return -1 if the destination is missing or so short as
to complete. This would of dropped us back onto the source with untidied references
in file engine and the cleanup route would be changed causing further errors.  The
only redeeming case is that it is hard to trigger without a specific setup