very small range requests on files were not handled fully
in cases like iphone which request ranges like 0-1 or similar small lengths, would cause
a client termination on file requests.  For the likes of 0-1 (iphone), auth is bypassed
and filled out as a flat buffer. Small ranges over 10 bytes get passed throught as normal
but the parser was not identifying the content and aborting the transfer.

If when parsing the buffer, all data is skipped over then treated as incomplete then only
error out if there is no range requested. With a range then there is an upper bound to
apply