add <redirect-to> tag to <mount>, keep backward compatability for <redirect>
because of previous behaviour of <redirect> introduce this tag which can take the form
of an exansion pattern (eg http://.../${mount}). The previous form of <redirect> is still
handled but as it automatically added the mount on the end then append that pattern onto
what is defined.  Obviously a static form is acceptable.