push ID3 tag detection into frame parser
ID3v2 tags were just filtered out at the file processing level but as ID3v2 can occur anywhere then
this treats it as a frame to match, which has to be careful as aac or mp3 etc could match.  The code
also allows for ID3v1 and APE but these are skipped when dealing with a stream.