small header updates, added missing date and maybe fix FF playback issue
after a report came in about FF playback, it seems that the accept-ranges header may be
causing some issue.  We don't need it and I'm not aware of anyone actually requiring it
so remove that for now.

Added Date header just to keep to standard, and only return 1.1 if the request was 1.1