fix for triggering auth remove event for missing request
Because the 404 route dropped the client onto the generic no_file handle, the auth lookup
failed to occur so auth remove triggers did not apply.  This makes a duplicate of that handle
(but does not place it in the internal tree) with a specific name so that the cleanup routine
does have a name.