block size scaling and listener throttling changes
The default max-send-size is set very high now. As block sizes are scaled based on
the incoming bitrate, there is no point in limiting the max write size as that will
just cause more syscalls.

The block sizing is tweaked to increase a little more on higher bitrates now (8mbs+,
up to about 16kper block), the per-listener send limiter has increased because of
this and the reschedule time on that limiter has also increased.  This allows for
larger writes while not going nuts on rescheduling.