update clients stats each second instead of each change
The stats may eventually only apply to if there is a change each second but for now on large
changes in the numbers of listeners, eg many streams or if there are many listeners on a single
stream, the clients stats is changing a lot.  This reduces the stats change work in such cases,
and reduces the amounts of information sent to stats clients.