fix possible crash case
odd one to trigger this. range request using auth which returns intro content being placed on an
inactive on-demand relay.  The range request can come from any html5 style player but would be a
small range but larger than 10 bytes and you would get an intro returned through auth.

The issue was that the intro flag was not getting reset when the intro content was replaced by the
small block. The routine to add a listener thought it was actual intro content, and allowed the
listener to procedd to kick off the inactive relay, but as it was not authenticated, ended up getting
dropped. The drop then had little to work on and ended up doing a NULL deref.