updates to the mp3/aac handler
very small intervals from relays were not handled properly, parsing metadata
itself worked but the subsequent offset calculation could become incorrect.

If not specified in the conf and the incoming interval is below 8k then use
the default interval, so that we don't do extra work. Very frequent metadata
updates is not that common but could be used in master to slave setups.

Some minor adjustments for block sizing