periodically check to see if the socket can be read, even if it has no content
It seems that with some proxies (eg aws), the read channel can be shut down while the write
channel can continue.  This state can occur if the client closes the link but the proxy keeps
the socket open.  This puts the socket into a close_wait state, ie the FIN is received but as
the general streaming routines are write only then it does not get detected.

This commit does a periodic check of the soket for reading, just a PEEK which detects an EOF
or unrecoverable error, dropping the client details.

if may be required later on to do this for keep-alive/1.1 only but for now at least it looks
sane and confirmed to work on aws.