small change to the cached entry pool handling. removed the global reference for removal
The ugly handling for removing entries from the cached tree is now gone. It only affected
the banned IP hander as that is the only one which is modified on an ongoing basis, the
others are based on config reloads.

This now includes a small pool of entries to note deletions so should allow for more
pruning when required.