tighten up processing when reloading ssl certs over reload
As we allow for updates to SSL certs over reload, there is a chance that the reload will fail
to get new details (eg typo in filename etc). Here we do not drop the existing SSL cert until
we have managed to load the new one (usually it's the same one of course).