add incoming client worker thread
Incoming clients processed by the listener thread pushes these clients to the normal worker
pool to do their processing. This isolates the IP filtering from the pool but other aspects
like xml processing and handling a large influx of clients (eg DDoS) were causing frequent
wakeups which in turn causes scans of the client list.

This commit intends to allow for a short term process worker that keeps them away from the
usual worker pool, so any initial work is performed allowing the balancing to move them to
the normal pool at a later stage.  For now the source based clients will migrate as well as
the in-memory clients (xml/xsl output).