fix log reopen bug if new file created
normally it is move/delete file and then HUP (logs reopen), but if someone created
a file (typically empty) before the HUP then we ended up not reopening.

we now add an extra check for the file size (actually verifying the name to an open
file is limited currently) to determine if the size is as expected. If on a reopen
case, we see a file present and is less than a trigger size then assume we are
appending and not archiving.