use avl tree on mount list for specific mount names
When large numbers of mounts are defined, it was taking a long time to do the recheck
routine (with lock held).  This was stalling certain tasks like registering a source
as the source tree lock was being held.  Now we place specific mounts in the avl for
quick lookup but still place wildcard mounts in the list as there needs to be some
priority order there.  The list is searched after the avl tree.