small API change for net to prevent silent fails to bind on sockets.
To allow for cases where people want to bind on all listening sockets, we have to
allow for the API to report such failures so that we can take action if we need to.

The API allows for a 0/-1 return to indicate whether there was a socket to tried, and
the actual socket in returned via a param, which could be SOCK_ERROR if it failed.
A failure is typically bind failures.  This could be expanded later so that the last
attempt can be described in a string of sorts.