avoid non-crash race case with updating source sent figures
a leftover from when the source was mutex locked, now it is only read locked from a
listener context. Here we be a little sneaky and use the rate handler to provide the
locking context, besides it is already taken in any case.

as the figures are informational, no crash is at issue, but the figures may be incorrect.