SSL cleanup and reduce global lock time held when accepting clients
with keepalive, the ssl part of the listener was not cleaned up so was wasting resources.
When accepting a connection, the global lock was held while creating the SSL context for the
new client. This re-arranges the setup so the reference for the listen socket is updated but
all other parts are done afterwards where the lock is not required.