make incoming rate to update far more frequently, but report the stat at the same rate
A number of internal checks are based on the current assessment of the incoming bitrate, which may
not be good enough if it is updated only every few seconds.  So update the internal value frequently,
eg every few read attempts, but report the stat for it at the same rate.