update internal metadata block layout for non-ogg streams.
Originally the only meta block was icy, so using a refbuf to store the icy data was fine but now we have
several blocks in parallel, eg FLV, iceblock.  To simplify the structure, we now define a layout to allow
for those to coexist without relying of generic type ordering.  The processing of the meta block data is
just the same, only the referencing is different.