fix possible crash issue with meta blocks
In cases where a copy is made from the queue, ie when on resync of lagging clients or a move to another
source and that block is not completely writen then a copy is made. This has traditionally been a block
copy with the blocks on the associated chain. With the metadata for mpeg streams change, the structure
was not a standard refbuf anymore so effectively copied the data using the wrong view.

The copy routine is now made format specific allowing for the proper interpretation with the default
being as-was, ogg still uses for layout. The detach routine is needed within introfile (the routine that
writes that incompleted blocks) and some some cleanup from legacy code in the copying itself.