enforce a range check in the queue sizing.
seems like the assessed range could get messed up because clients could drop out and
any time and leave the targeting too short.  You could also get a lagging client
also stretching the queue but that is less of an issue as that tends to resolve on a
later interation.  The former though could cause excessive purging.