avoid generic reader reschedule if block has been added. and clean up silly log message
silly log message was from the intro file load when there isn't one to load

the skip setting was adversely affecting the reader part of certain high bandwidth streams. These are
typically Mb/s. The rescheduling ended up increasing in certain cases to a point where lowering it
would take it so long that the source would drop.  This has shown when the source does not send anything
straight away then starts blasting the content