prevent incorrectly defined mime types from breaking relays
This is not usually a problem, but if one is define and it is not a recognised type then some
initialisation of relays can be missed out on retrying.  Initially when relays a created they
are setup and work fine, but on retry, the structures and settings are not cleared out as type
checking for listeners is maintained but initialisation needs to be done for certain things
like metadata inserts.

This mismatch was tripping a bug here causing metadata alignment problems. Now we verify the
specified type before overriding.