2 fixes, set a default burst/min queue size for source, and allow response on failed relay
Using the seconds measure for some of these sizes works but as the assessment is every few seconds then
it may take a little time to get up to speed, which is an issue for on-demand relays. Here we provide a
default until then (only a few seconds).

The routine for waiting listeners (for on-demand relaying to start up) had an issue were it would flag
as error on relay failure, meaning no 404 type response.  We now avoid that case and exit as per normal
and issue a 404 response.