update for glibc api on threading.
as we use rwlocks for source handling (listeners only read), it can lead to writer
starvation in heavily loaded cases. rwlocks by default are set to prefer readers but
I had non-portable code for setting prefer writers (glibc only) but the API aspect
had changed and defaulted back to preferring readers. This allows a change on a certain
assumption which is valid for icecast.