rework previous fix for pthread rwlock preferences
seems that the code for setting writers as preference was not working as the settings were enums which
messed up the #ifdef attribs. This simplifies the code by using the initializer macro whih if not present
reverts back to the normal init without attr.  This means its the default unless overridden. on glibc it
seems to be readers, on bsd it seems to be writers, winpthreads is unclear.