crash fix case with using stream auth and metadata updates.
This stems from a cleanup of admin clients. clients passed to auth threads are only marked
inactive to avoid workers looking at them, but admin mount requests also want to move to the worker
for the source however the auth thread cannot do that without leaving stale data in the worker the
client resides on.  This has to be done by the worker itself, so reactivate the client and retry to
trigger a move.