update for triggers for the listeners moving between workers
Avoid a caching issue, and be more fine grained in when to check for certain settings. In the case
of the incoming worker, the change should happen quickly but for normal workers, if we skip the odd
time it is not an issue and saves on locking.

As the slave thread triggers the least busy check every second then the listener does not, which is
where the locking might of fallen down. And now the listener routine bypassess if the check cannot
be done, saves blocking the worker.

On testing, the distribution of listeners is working much better and in a responsive way