pull out the add block code intoa separate function so it can be a callback or used as-is
normally the process involves read, return block, add block, back to worker. But there are cases where
it would be better to have the read routine call the add directly instead of returning the block.  The inital
example of this would be where keyframe blocks are used.

In the case of keyframes, you want listeners to find them quickly which may be an issue if the spacing between
the keyframes is larger than the min/burst sizes. In those case the listener routines will rescan many times
before finding it.

This allows the reader routine to possible cache the blocks until it is ready to put them up for listeners.