Add extra log lines and avoid extra auth url work in extreme cases
Just expand on the line limit checks for the non-add listener requests
and kick out before attempting the auth. This does not show as overflow
but if anyone does try to spam then it does limit the auth attempts.

Should only be a factor on listener remove, mount add/remove and stream
auth.