make auth_client have a flags setting to mirror the client flags
As the worker does refer to client flags member (even inactive clients)
then to avoid having to take the lock, we should mirror it in the auth
client details and transfer it after processing.

Eventually we should treat it as a worker change client and remove it
from the workers onto auth only to process and then re-add it. Then this
commit for mirror flags can be removed.