http header update
this covers the majority of the use cases WRT clients. Drop the var args
part for the complete routine and just use the separate call for any refbuf
list.

flv wraps the initial header within the write routine not the client data
setup.

The general headers routine is renamed although does the same thing using
the newer http routines.