Multiple relay host switchover handling
While cascading down hosts has worked for some time, resuming a failed
host has not been implemented automatically. You could of done a
killsource to trigger it but it was not done automatically.  This
feature now looks to be working.

We don't want to interrupt the existing stream until we know we have a
working replacement, so we kick off a parallel thread with a relay copy
to establish a working client. If/when established, we then need to
switch in the feed to the source.  As we already have a client handle
and a copy of the original relay, we lookup the existing source and flag
it to tell the existing relay handler to detach itself allowing the new
one to latch on.

The process involved a new source flag for switchover (might be useful
in other situations) and wait until it drops (done by the existing relay
handler to indicate that it has now let go of it). At that point the new
relay/client attaches to the stream and the feed resumes. The listeners
are unaffected from the swap over (content aside)