ninor changes to relay copy and setup
This should have no external effect on things as the copy routine is just
copying a few more details of internal state, and as it is copied from
either xml or master then there is no state to talk of. The in_use member
is just set later when the stream is initialised, it is still used.