move icy state data to the source client handle instead of the format
This is more for later work. If a client feeding a source is changed then
the state of the metadata alignment is not relevant, so keep that tied to
the client and use the usual cleanup routines to flush that out.

Most of the commit is just relating to the name changes. Using icy instead
of mp3 as it is not just mp3 anyore.  The interval code needed more
alteration though.