relay hijack coding
Still modification as it has not been simplified yet, but the basics are
there. Using a similar process to the higher host priority switchover, a
source can jump in to override an existing relay or an existing source
client.  In the relay case, we need to drop off the client but leave the
relay ready to go. On the source client exit, that relay is searched for
and a new client is added to get it going.