update xml parsing routine to run tag as parameters
The original version of this worked well for normal xml tag for the
settings but sometimes people want data as parameters eg

<mount mount-name="/stream">

Also a number of settings are now being added as parameters, alias, auth
http-headers etc.  This update allows for the tables to used for tags
and parameters. parameters are obviously for simpler types so mount is
not going to be a parameter but the default is to allow but a flag can
be coded on the option to not check it as a parameter.

This has also led to pulling in the authentication parsing block from
auth into cfgfile which was an oddity.

The actual processing/storage of actual items is no different just a
small change to the API