add priority setting on mount block
This is used for the ordering on the wildcard list, not the avl tree. The
avl tree stores explicitly name mounts which are searched first so have
highest priority for lookups and are unique.

In a missing explicit mount, a requested mountpoint can match multiple
pattern matching templates. Normally it matches the last one listed in the
xml but you may want to define one general, and others that override but
not care too much about the ordering.

The priority setting if missing defaults to the lowest priority, and the
search looks for the last one in the highest priority grouping. This
should allow for compatability to existing xml configs but allow for admins
more flexability in setup.