add wrapper for curl handle for exit handling
To have a cleaner exit path for request going via curl handles, use the
callback routines (used for progress reports) to check the status of the
servers running state. If it is shutting down then do not immediately abort
those transfers, but do after a number of seconds have elapsed.

While an immediate abort would stop quicker with in-flight requests, by
allowing a certain duration to elapse means allowances for tracking state
exterally is kept.