extend on http work to handle POST options.
The http part is essentially the same, the difference is in relation to
dividers and end of pariing characters.  The http complete also explicitly
adds a blank line because the complete routine does not add the end of
paring characters to the last entry.  This waym for http you get the blank
with nothing following it but the previous has a \r\n, in the POST case
there is nothing added on the end so no & is at the end.

The POST variants don't refer to a client refbuf, instead it returns one to
allow the caller to decide what to do. auth url does not use a client for
example. YP does not either.