Use POST fields builder and rework some of the callback routine
a chunk of this is the conversion over to using the http work, which builds
the post fields which are passed into the libcurl handle.

The callback routines for headers and data were not that great. Most of it
stems from the problem that headers are not required to be nul-terminated,
but it's not clear on whether the buffer passed is really changeable. So
to be clearer on this, and to make processing more in-line with strings
with nuls, these routines were re-worked to do the same sort of thing be
to make sure there is not gotchas