set the initial time correctly for rechecking hosts
when the relay is established, set the recheck hosts initial time else
it will retry immediately.  Most likely not something of concern just
some extra processing work at the beginning but the feed may sound like
a small burst of audio at switchover if the higher priority feed just
drops for a moment.