minor relay related cleanups
allow a larger range of relays to start up. not excessive but allows more
in case many are configured.

introduce a start member for relay so that we can process a the startup
routine without kicking off a relay early. This is targetted at relays
that are offline but we may need to do internal housekeeping. not used
much for now.