introduce the linger-for mount tag
This is to specify the number of seconds a terminating stream stays active
for before it finally shuts down. This has a few possibilities for later
commits but if a relay terminates, you may want to delay the reconnection
by a short time to allow the sources to reconnect upstream.  The basic
idea is the the stream put into a state where no further stream reading is
performed but all other actions are.

The main change in this commit is the splitting of the source_read routine
into 2, one being a wrapper for catching the returns from the internal one

The linger facility will also be able to interact with the switchover
mechanism.