a small tweak for header setting and checking on relays
When sending the Host header, avoid putting on the port if it is non-ssl on
port 80.   It should not really be needed but one report had an issue which is
to be fixed.

If the response comes back with chunked data then drop it for now. It could be
handled but it's unclear on how that interacts with icy-metadata.

karl