avoid relay initialize failure case
When the source reserve routine became non-blocking, one case started causing a
possible issue. If the source tree could not get a lock then the return was
similar to a failed to allocate source (eg already in use).   Here we will
require the source tree lock to be acquired before entering the function so
that the exit path is simplified.

For relays, the relay_has_source is the wrapper routine and works handles that
aspect, but we do not need it for the update_relays routine as it returns in
such cases to do the initialise from the worker thread.