fix missing state info in relay on xml reload
on reload, the relays are copied into place but do not have the current state
information that the active one do. Instead of complicating the copy routine
with a third relay for state information, use the get relay_details routine that
will have access to the current active one and the newly copied one.  When
changeover is detected, we can traverse the hosts at that point and update the
state info on the new relay details.  A comparison of the ip/port/mount between
the two and act as a trigger to do a relay restart else leave as-is.