allow for setting some extra flags in logging
Not normally used. The simple one here is to expand the timestamps to include
microseconds by setting flag on the level.

bit 8 set for enable microseconds on logging. This means that 256 + 4 means
debug level with microseconds.

The 8 bits is to allow for larger range of other levels to be defined, also
other flags may also be possible.