update on worker processing of clients
In general, the process is the same but the internal API has changed.
We now have a queue for fast response clients and clients that have a
short schedule time are placed on there.  Clients will often bounce
between the normal and fast queues as required so the ordering of
clients will change.

The count represents the clients on both normal and fast queues but the
pending queue is still counted separately.

Certain members of the worker were removed to a control structure to
separate access, the clients do not get reference to those.