rework for flac streams
rework internals to better handle the report-0 header case. It's unclear if this
is that common but to avoid the possibility introduce a variable header facility
and handle a case of those packets being larger than a page.

mpd seems to send a report-0 headers then resets to send report-1 header.  Now
the code path has a header and non-header phase, and during the header phase,
each page is pagein and packets pulled out and their type is then checked. On
first detection of non-header, then the phase is switched.  This avoids the
relying on the header count as 0 is classed as valid.

The cached pages can now be used to keep pages that may not be known for use as
headers as you do not know that until all of the packet is pulled out.  If there
are any pages that are in that state (may never happen) then they get added to
the actual headers