log internal change.
2 things changed here. procedural aspects and locking are unchanged

The expansion from timestamp to localtime/strftime is pushed back to the
logging thread, allowing for quicker return to worker threads. The timestamp
is still acquired at the front end for accuracy.

All priority messages are now passed to the logging thread and the thread
filters the priority messages.

The former reduces the locking range in the workers but the latter adds work
for low priority settings.  This allows for debug messages to be caught in the
internal log in core dump cases, we can also re-introduce the priority filter
at the front end at a later point.