split locking, caching log entries and small API updates
global rwlock used instead of mutex, read lock mostly used to freeze external
caller settings. a mutex is used internally to protect queue integrity. half of
the log_t structure is only used by the log thread.

The close trigger is done via a setting a flag that the log thread catches.

The log line has a reserved space for a log line value, allows for identifying
lines in a core file, no other use currently.  Each log entry has space for a
timestamp which can be expanded to local time by the log thread.  At the moment
all log entry are sent to the log thread no matter the level setting, allowing
for extracting entries even if they did not get to the log file.

The priorities part is the more involved part. As the higher level log messages
have larger volume throughput, a simple queue for the history has a problem as
the lower level messages will get pushed out very quickly. To mitigate this we
keep a simple queue for the logging in time out, but each message is placed on
a queue for each priority and so many is kept on each priority.
This will make the log look different, as you keep N messages of each priority
but it is in time order.

To mitigate the number of log thread triggers with higher volume, the wake up
trigger is moved from the add to queue to a regular schedule from the slave
thread. This will then treat as a regular push instead of an on-demand trigger

I may change the keep count for the priorities, but that will require a parsing
change in the xml. TBD