log level code update
This is to help clear up some of the interaction for the priority handling in
the log priority handling.  While not 100% configurable, it is getting there.

The logging code now provides the mapping for the names and how many messages
are actually kept. To facilitate this a simpler structure is used for external
to the log code that can provide certain configuration info. This separates the
basic info from the internals for processing the queues.

The log_levels_t structure is there to provide the basic info without all the
internal detail. A few access routines to interact and convert the _kept call
turned into a wrapper for the replacement routine.  For access log etc, there
are no priorities as such so the wrapper is sufficient.

eventually, the idea from pulling out the format for error log lines into
logging should be considered.  Important if the layout may need to be updated

The level xml tag entry is now more involed. It still takes the traditional
number 1-4 and the more recent symbolic form (info etc) but now, to allow for
the different length stored and represented, a more involved form is possible

<level>N</level> where N can be a comma separated list

0-4   The traditional levels

error,warn,info,debug   The more recent symbols for the above

subsec    For logging subsecond figures, more a debug thing

name=value  Where name can be one error,warn,info,debug and value being the
number of messages kept for that priority in the error log