auth update. messaging updates and separate timers for url actions
the timeout handler was shared between all the routines in auth, so a failure on
a listener remove could stop listener add from running.  These are now separated
and set to independently restart.  This was primarily an issue between listener
adding and removing.

I've also cleaned up some of the messaging to help diagnostics.