rearrangement of options reading in url auth. delayed auth warning message
this allows for the timeout and on_error_wait settings to be set at any point
in the listing.  Also, issue a warning message if the url request suceeds but
is taking over 3seconds. A sporadic delay is not critical but a continual
series of delays can be more of a cause.  Better to log it than ignore.