expand on the relay host skip setting, convert to a timestamp
To avoid poor situation where a higher priority relay host kicks in but only
lasts a very short time, treat any short lasting stream as something not to be
retried as frequently as normal down connection.  A normal outage can be tried
every few seconds where as the stream switch to a feed that dies is more of a
problem we have to work around.  It may be necessary to somehow disable it to
avoid it but for now, expand on the delay time.