update to range handling.
here we keep a start and end position in the connection_t and use that to cause
a stream stop.  we take into account the difference on files and streams, and
if there is anything to skip over at the beginning of a file.

The oddball here is safari with its multiple requests to assess things, but
hopefully that can be tightened up to sort it.