pick out the web/admin changes to drop the rounded corners from pull #242
Use the css aspects to use the in-built browser use which allows us to avoid
the extra requests to get the images.

The other part of the pull is the multi-user shoutcast compatible handling. That
already exists using a similar type of layout but whether there is something
different in the implementation will need to assessed.  The web page changes
though are fine.

karl.